4.21 exercise completed

+++ part4/bloglist/controllers/blogs.js
@@ -73,8 +73,30 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const blog = await Blog.findById(request.params.id)
+
+  if (!blog) {
+    return response.status(404).json({ error: 'Blog not found' })
+  }
+
+  if (!request.token) {
+    return response.status(401).json({ error: 'token missing' })
+  }
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'Invalid token' })
+  }
+
+  if (blog.user && blog.user.toString() === decodedToken.id) {
+    await Blog.findByIdAndDelete(request.params.id)
+    console.log('Blog was deleted')
+    return response.status(204).end()
+  } else {
+    return response.status(403).json({
+      error: 'Permission denied, you can delete only your own blogs'
+    })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
